Finish gpt-2 and gpt-j API's
