+++ examples/gpt-2/CMakeLists.txt
@@ -2,8 +2,10 @@
 # gpt-2
 
 set(TEST_TARGET gpt-2)
-add_executable(${TEST_TARGET} main.cpp)
+add_executable(${TEST_TARGET} gpt-2.cpp main.cpp)
 target_link_libraries(${TEST_TARGET} PRIVATE ggml ggml_utils)
+target_include_directories(${TEST_TARGET} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+
 
 #
 # gpt-2-quantize
